requirements.txt
fastapi
uvicorn
kaggle
pandas
numpy
pillow
tqdm
imagehash
albumentations==1.4.8
timm==1.0.8
torch
torchvision
scikit-learn
matplotlib
onnx
onnxruntime
pyyaml
python-multipart
httpx
opencv-python
streamlit
